+++ src/seekstorm/add_result.rs
@@ -3221,7 +3221,6 @@ pub(crate) fn add_result_multiterm_singlefield(
                 Ordering::Equal => {
                     if t2 + 1 < non_unique_query_list.len() {
                         t2 += 1;
-
                         pos2 = non_unique_query_list[t2].pos;
                         continue;
                     }