+++ src/seekstorm/lib.rs
@@ -40,7 +40,7 @@
 //! };
 //! let segment_number_bits1=11;
 //! let serialize_schema=true;
-//! let index=create_index(index_path,meta,&schema,serialize_schema,segment_number_bits1).unwrap();
+//! let index=create_index(index_path,meta,&schema,serialize_schema,&Vec::new(),segment_number_bits1).unwrap();
 //! let _index_arc = Arc::new(RwLock::new(index));
 //! ```
 //! ### open index (alternatively to create index)
@@ -113,7 +113,7 @@
 //!     highlight_markup: true,
 //! },
 //! ];    
-//! let highlighter=Some(highlighter(highlights, result_object.query_term_strings));
+//! let highlighter=Some(highlighter(&index_arc,highlights, result_object.query_term_strings));
 //! let return_fields_filter= HashSet::new();
 //! let mut index=index_arc.write().await;
 //! for result in result_object.results.iter() {
@@ -167,7 +167,6 @@
 //!   }
 //! ]
 //! ```
-//!
 //! ```rust
 //! let file_path=Path::new("C:/Users/johndoe/Downloads");
 //! let _ =index_arc.ingest_pdf(file_path).await;
@@ -256,7 +255,7 @@
 //! };
 //! let serialize_schema=true;
 //! let segment_number_bits1=11;
-//! let index=create_index(index_path,meta,&schema,serialize_schema,segment_number_bits1,false).unwrap();
+//! let index=create_index(index_path,meta,&schema,serialize_schema,&Vec::new(),segment_number_bits1,false).unwrap();
 //! let mut index_arc = Arc::new(RwLock::new(index));
 //! ```
 //! ### index documents
@@ -297,7 +296,7 @@
 //!             highlight_markup: true,
 //!         },
 //!     ];    
-//! let highlighter2=Some(highlighter(highlights, result_object.query_terms));
+//! let highlighter2=Some(highlighter(&index_arc,highlights, result_object.query_terms));
 //! let return_fields_filter= HashSet::new();
 //! let index=index_arc.write().await;
 //! for result in result_object.results.iter() {